Changed affected transcript overlap percentage property type.

+++ Unite.Indices/Entities/Basic/Genome/Variants/AffectedTranscriptIndex.cs
@@ -9,5 +9,5 @@ public class AffectedTranscriptIndex
 
     public int? Distance { get; set; }
     public int? OverlapBpNumber { get; set; }
-    public int? OverlapPercentage { get; set; }
+    public double? OverlapPercentage { get; set; }
 }
